+++ src/main/java/ru/vyarus/guice/persist/orient/repository/command/ext/listen/ListenParamExtension.java
@@ -2,14 +2,15 @@ package ru.vyarus.guice.persist.orient.repository.command.ext.listen;
 
 import com.orientechnologies.orient.core.command.OCommandRequest;
 import com.orientechnologies.orient.core.command.OCommandRequestAbstract;
-import com.orientechnologies.orient.core.command.OCommandResultListener;
+import ru.vyarus.guice.persist.orient.db.util.Order;
 import ru.vyarus.guice.persist.orient.repository.command.core.param.CommandParamsContext;
 import ru.vyarus.guice.persist.orient.repository.command.core.spi.CommandExtension;
 import ru.vyarus.guice.persist.orient.repository.command.core.spi.CommandMethodDescriptor;
 import ru.vyarus.guice.persist.orient.repository.command.core.spi.SqlCommandDescriptor;
+import ru.vyarus.guice.persist.orient.repository.command.ext.listen.type.LiveListenerTypeSupport;
+import ru.vyarus.guice.persist.orient.repository.command.ext.listen.type.QueryListenerTypeSupport;
 import ru.vyarus.guice.persist.orient.repository.core.spi.parameter.MethodParamExtension;
 import ru.vyarus.guice.persist.orient.repository.core.spi.parameter.ParamInfo;
-import ru.vyarus.guice.persist.orient.db.util.Order;
 
 import javax.inject.Singleton;
 import java.util.List;
@@ -32,19 +33,22 @@ public class ListenParamExtension implements
 
     public static final String KEY = ListenParamExtension.class.getName();
 
+    private final ListenerTypeSupport queryHandler = new QueryListenerTypeSupport();
+    private final ListenerTypeSupport liveHandler = new LiveListenerTypeSupport();
+
     @Override
+    @SuppressWarnings("unchecked")
     public void processParameters(final CommandMethodDescriptor descriptor, final CommandParamsContext context,
                                   final List<ParamInfo<Listen>> paramsInfo) {
         check(paramsInfo.size() == 1, "Duplicate @%s definition", Listen.class.getSimpleName());
-        check(descriptor.command.startsWith("select"), "Listener could be applied only for select queries");
-        final Class<?> returnType = context.getDescriptorContext().method.getReturnType();
-        check(void.class.equals(returnType) || Void.class.equals(returnType),
-                "Method with listener must be void, because no results returned from query "
-                        + "when listener used");
         final ParamInfo<Listen> param = paramsInfo.get(0);
-        check(OCommandResultListener.class.isAssignableFrom(param.type), "Only %s can be used as result listener",
-                OCommandResultListener.class.getName());
-        descriptor.extDescriptors.put(KEY, param.position);
+        final Class<?> returnType = context.getDescriptorContext().method.getReturnType();
+        final String query = descriptor.command.toLowerCase();
+        final ListenParamDescriptor extDesc = new ListenParamDescriptor(selectHandler(query), param.position);
+
+        extDesc.handler.checkParameter(query, param, returnType);
+
+        descriptor.extDescriptors.put(KEY, extDesc);
     }
 
     @Override
@@ -54,15 +58,44 @@ public class ListenParamExtension implements
     }
 
     @Override
+    @SuppressWarnings("unchecked")
     public void amendCommand(final OCommandRequest query, final CommandMethodDescriptor descriptor,
                              final Object instance, final Object... arguments) {
-        checkExec(query instanceof OCommandRequestAbstract,
-                "@%s can't be applied to query, because command object %s doesn't support it",
-                Listen.class.getSimpleName(), query.getClass().getName());
-        final Integer position = (Integer) descriptor.extDescriptors.get(KEY);
-        final OCommandResultListener listener = (OCommandResultListener) arguments[position];
+        final ListenParamDescriptor extDesc = (ListenParamDescriptor) descriptor
+                .extDescriptors.get(ListenParamExtension.KEY);
+        final Object listener = arguments[extDesc.position];
         // null listener makes no sense: method is void and results are not handled anywhere
         checkExec(listener != null, "Listener can't be null");
-        ((OCommandRequestAbstract) query).setResultListener(listener);
+
+        ((OCommandRequestAbstract) query).setResultListener(extDesc.handler.processListener(query, listener));
+    }
+
+    private ListenerTypeSupport selectHandler(final String query) {
+        return query.startsWith("live") ? liveHandler : queryHandler;
+    }
+
+    /**
+     * Extension internal model.
+     *
+     * @author Vyacheslav Rusakov
+     * @since 29.09.2017
+     */
+    @SuppressWarnings("checkstyle:VisibilityModifier")
+    public static class ListenParamDescriptor {
+
+        /**
+         * Selected listener type handler.
+         */
+        public final ListenerTypeSupport handler;
+
+        /**
+         * Listener parameter position.
+         */
+        public final int position;
+
+        ListenParamDescriptor(final ListenerTypeSupport handler, final int position) {
+            this.handler = handler;
+            this.position = position;
+        }
     }
 }